+++ apps/server/src/modules/h5p-editor/controller/dto/ajax/post.body.params.ts
@@ -1,16 +1,14 @@
-import { Injectable, PipeTransform, ValidationPipe } from '@nestjs/common';
 import { ApiProperty } from '@nestjs/swagger';
-import { plainToClass } from 'class-transformer';
-import { IsArray, IsMongoId, IsOptional, IsString, validate } from 'class-validator';
+import { IsArray, IsMongoId, IsOptional, IsString } from 'class-validator';
 
-class LibrariesBodyParams {
+export class LibrariesBodyParams {
 	@ApiProperty()
 	@IsArray()
 	@IsString({ each: true })
 	libraries!: string[];
 }
 
-class ContentBodyParams {
+export class ContentBodyParams {
 	@ApiProperty()
 	@IsMongoId()
 	contentId!: string;
@@ -21,44 +19,10 @@ class ContentBodyParams {
 	field!: string;
 }
 
-class LibraryParametersBodyParams {
+export class LibraryParametersBodyParams {
 	@ApiProperty()
 	@IsString()
 	libraryParameters!: string;
 }
 
 export type AjaxPostBodyParams = LibrariesBodyParams | ContentBodyParams | LibraryParametersBodyParams | undefined;
-
-/**
- * This transform pipe allows nest to validate the incoming request.
- * Since H5P does sent bodies with different shapes, this custom ValidationPipe makes sure the different cases are correctly validated.
- */
-@Injectable()
-export class AjaxPostBodyParamsTransformPipe implements PipeTransform {
-	async transform(value: AjaxPostBodyParams) {
-		if (value) {
-			let transformed: Exclude<AjaxPostBodyParams, undefined>;
-
-			if ('libraries' in value) {
-				transformed = plainToClass(LibrariesBodyParams, value);
-			} else if ('contentId' in value) {
-				transformed = plainToClass(ContentBodyParams, value);
-			} else if ('libraryParameters' in value) {
-				transformed = plainToClass(LibraryParametersBodyParams, value);
-			} else {
-				return undefined;
-			}
-
-			const validationResult = await validate(transformed);
-			if (validationResult.length > 0) {
-				const validationPipe = new ValidationPipe();
-				const exceptionFactory = validationPipe.createExceptionFactory();
-				throw exceptionFactory(validationResult);
-			}
-
-			return transformed;
-		}
-
-		return undefined;
-	}
-}
